stupid broken DMA driver
I couldn't get interrupts to work, so it just busy-waits, and it only works 50% of the time for some reason - like, reliably. If it worked this time, it won't work next time, and if it didn't work this time, it's guaranteed to work next time.

I'm probably not going to keep working on this - if I'm going to busy-wait, I may as well just use CSRs which have been infinitely more reliable this entire time. I don't need maximum performance.